fix: 2 bugs with collapsible block state and properties
When 2 db graphs exist, the built-in-properties check failed because it
failed to db lookup for the new repo. Just simplified it so that db
lookup is only needed once across repos. Also limited check to hidden
built-in properties since visible built-in properties should be
treated like user properties and be collapsable