enhance: improve :block/tags validation
The move from :block/type to :block/tags allows for a couple
types of bugs that weren't possible before. Add validations
to ensure built-in entities aren't being accidentally modified
by the UI and that :logseq.class/Page remains simple and a useful
tag for features later. The Page validation is turned off until an
importer bug is fixed